Fix calls to the settings API
Most of POST to the settings API changed to either PUT or PATCH.
Please notice the sleep between each test. This is to give some time for meilisearch.
Without it, the tests sometimes fail when started collectively. But when executed
individually each test succeeds.

+++ src/test/java/com/meilisearch/integration/classes/AbstractIT.java
@@ -46,7 +46,11 @@ public abstract class AbstractIT {
     }
 
     public static void cleanup() {
-        deleteAllIndexes();
+        cleanup(null);
+    }
+
+    public static void cleanup(String prefix) {
+        deleteAllIndexes(prefix);
         apacheHttpClient.shutdown();
     }
 
@@ -101,12 +105,16 @@ public abstract class AbstractIT {
         return null;
     }
 
-    public static void deleteAllIndexes() {
+    public static void deleteAllIndexes(String prefix) {
         try {
             Client ms = new Client(new Config(getMeilisearchHost(), "masterKey"), 
                                                 new MeiliSearchHttpRequest(apacheHttpClient));
             Index[] indexes = ms.getIndexList();
             for (Index index : indexes) {
+                String indexUid = index.getUid();
+                if (prefix != null && !indexUid.startsWith(prefix)) {
+                    continue;
+                }
                 ms.deleteIndex(index.getUid());
             }
         } catch (Exception e) {
